Extract interface AsyncApiDocketService
